+++ gns3server/agent/mcp/drawings.py
