+++ gns3server/agent/mcp/images.py
