+++ gns3server/agent/mcp/snapshots.py
