+++ gns3server/agent/mcp/symbols.py
