+++ gns3server/agent/mcp/templates.py
